+++ app/Http/Controllers/API/TagController.php
@@ -2,13 +2,16 @@
 
 namespace App\Http\Controllers\API;
 
-use App\Http\Controllers\Controller;
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 use App\Http\Resources\Tag as TagResource;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class TagController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -24,6 +27,7 @@ class TagController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Tag::class);
         $tag = new TagResource(Tag::create($request->all()));
         return $tag->response()->setStatusCode(200);
     }
@@ -43,6 +47,8 @@ class TagController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $idtag = Tag::findOrFail($id);
+        $this->authorize('update', $idtag);
         $tag = new TagResource(Tag::findOrFail($id));
         $tag->update($request->all());
 
@@ -54,6 +60,8 @@ class TagController extends Controller
      */
     public function destroy($id)
     {
+        $idtag = Tag::findOrFail($id);
+        $this->authorize('delete', $idtag);
         Tag::findOrFail($id)->delete();
         return 204;
     }
